refactor milestone tests and address comments
