+++ src/express/commands/milestone-partition.js
@@ -9,54 +9,31 @@ import {
   getEnode,
   validateFinalizedBlock,
   fetchLatestMilestone,
-  joinAllPeers
+  joinAllPeers,
+  getUsersAndHosts,
+  getIpsAndEnode,
+  fetchAndValidateSameBlocks
 } from '../common/milestone-utils'
 
 const { maxRetries, runCommand } = require('../common/remote-worker')
 
-const milestoneLength = 64
+const milestoneLength = 12
 const queryTimer = (milestoneLength / 8) * 1000
 
 export async function milestonePartition() {
-  // NOTE: Make sure bor branch has logic for hardcoded primary validator
-
-  require('dotenv').config({ path: `${process.cwd()}/.env` })
-  const devnetType =
-    process.env.TF_VAR_DOCKERIZED === 'yes' ? 'docker' : 'remote'
-
-  const doc = await loadDevnetConfig(devnetType)
-
-  const borUsers = splitToArray(doc['devnetBorUsers'].toString())
-  const borHosts = splitToArray(doc['devnetBorHosts'].toString())
+  // Get users and hosts
+  let borUsers, borHosts = await getUsersAndHosts
 
   // Check for number of validators
-  if (doc['numOfValidators'].length < 4) {
+  if (borUsers.length < 4) {
     console.log('📍Cannot run milestone tests on less than 4 validator nodes')
     process.exit(1)
   }
 
-  // Grab the enode of all the nodes
-  let enodes = []
-  let tasks = []
-  const ips = []
-  for (let i = 0; i < borUsers.length; i++) {
-    const ip = `${borUsers[i]}@${borHosts[i]}`
-    ips.push(ip)
-    tasks.push(getEnode(borUsers[i], borHosts[i]))
-  }
-
-  await Promise.all(tasks).then((values) => {
-    enodes = values
-  })
-
-  if (enodes.includes('')) {
-    console.log('📍Unable to fetch enode, exiting')
-    return
-  }
+  // Get IPs and enodes of all nodes
+  let ips, enodes = await getIpsAndEnode(borUsers, borHosts)
 
   console.log('📍Rejoining clusters before performing tests')
-
-  // Make sure all peers are joined
   let joined = await joinAllPeers(ips, enodes)
   if (!joined) {
     console.log('📍Unable to join peers before starting tests, exiting')
@@ -76,109 +53,30 @@ export async function milestonePartition() {
     return
   }
 
-  console.log('📍Waiting 10s to fetch finalized blocks...')
+  console.log('📍Waiting 10s to to fetch and validate finalized blocks...')
   await timer(10000)
 
-  // Fetch the last 'finalized' block
-  console.log('📍Trying to fetch last finalized block')
-  const finalizedBlock = await runCommand(
-    getBlock,
-    borHosts[0],
-    'finalized',
-    maxRetries
-  )
-  if (finalizedBlock === undefined) {
-    console.log('📍Unable to fetch last finalized block, exiting')
-    return
-  }
-
-  // Check if the number and hash matches with the last milestone
-  if (
-    Number(finalizedBlock.number) === Number(lastMilestone.end_block) &&
-    finalizedBlock.hash === lastMilestone.hash
-  ) {
-    console.log(
-      '📍Received correct finalized block according to last milestone'
-    )
-  } else {
-    console.log(
-      `📍Block number or hash mismatch for finalized block. Finalized Block Number: ${Number(
-        finalizedBlock.number
-      )}, Hash: ${finalizedBlock.hash}. Milestone end block: ${Number(
-        lastMilestone.end_block
-      )}, Hash: ${lastMilestone.hash} exiting`
-    )
-    return
-  }
+  // Validate the 'finalized' block with last milestone
+  await validateFinalizedBlock(borHosts, lastMilestone)
 
   console.log('📍Creating clusters for tests')
 
   // Next step is to create 2 clusters where primary node is separated from the
-  // rest of the network.
-  let created = await createClusters(ips, enodes, 2)
-  if (!created) {
-    console.log('📍Unable to remove peers for creating clusters, exiting')
-    return
+  // rest of the network. For a partition based test case, the split will be 50:50
+  // i.e. out of 4 nodes equal partition of 2-2 nodes will be created. 
+  await createClusters(ips, enodes, 2)
+  if (!valid) {
+    console.log(`📍Failed to create partition clusters, retrying`)
+    valid = await createClusters(ips, enodes, 2)
+    if (!valid) {
+      console.log(`📍Failed to create partition clusters, exiting`)
+      process.exit(1)
+    }
   }
 
-  // Validate if the cluster is created by number of peers
-  tasks = []
-  for (let i = 0; i < ips.length; i++) {
-    tasks.push(getPeerLength(ips[i]))
-  }
-
-  let peers = []
-  await Promise.all(tasks).then((values) => {
-    // Check if there's no validation error
-    if (values.includes(-1)) {
-      console.log('📍Unable to query peer length, exiting')
-      return
-    }
-    peers = values
-  })
-
-  // validate if number of peers are correct or not
-  const expected = [1, 1, 1, 1]
-  if (JSON.stringify(peers) !== JSON.stringify(expected)) {
-    console.log(
-      `📍Retrying creation of partition clusters for testing due to peer length mismatch, got: ${peers}, expected: ${expected}`
-    )
-    created = await createClusters(ips, enodes, 2)
-    if (!created) {
-      console.log('📍Unable to remove peers for creating clusters, exiting')
-      return
-    }
-
-    // Validate if the cluster is created by number of peers
-    tasks = []
-    for (let i = 0; i < ips.length; i++) {
-      tasks.push(getPeerLength(ips[i]))
-    }
-
-    peers = []
-    await Promise.all(tasks).then((values) => {
-      // Check if there's no validation error
-      if (values.includes(-1)) {
-        console.log('📍Unable to query peer length, exiting')
-        return
-      }
-      peers = values
-    })
-
-    if (JSON.stringify(peers) !== JSON.stringify(expected)) {
-      console.log(
-        `📍Peer length mismatch while creating clusters, got: ${peers}, expected: ${expected}`
-      )
-      console.log('📍Failed to create partition clusters for testing, exiting')
-      return
-    } else {
-      console.log(
-        '📍Partition clusters for testing created. Proceeding to test'
-      )
-    }
-  } else {
-    console.log('📍Partition clusters for testing created. Proceeding to test')
-  }
+  console.log(
+    '📍Partition clusters for testing created. Proceeding to test'
+  )
 
   // Reaching this step means that we've created 2 clusters for testing.
   // Cluster 1 has 2 nodes with 1 primary producer whose difficulty will always be higher.
@@ -189,74 +87,9 @@ export async function milestonePartition() {
   console.log('📍Waiting 10s before fetching latest block from both clusters')
   await timer(10000)
 
-  // We'll fetch block from cluster 2 first as it'll be behind in terms of block height
-  let latestBlockCluster1, latestBlockCluster2
-  latestBlockCluster2 = await runCommand(
-    getBlock,
-    borHosts[2],
-    'latest',
-    maxRetries
-  )
-  if (latestBlockCluster2 === undefined) {
-    console.log('📍Unable to fetch latest block in cluster 2, exiting')
-    return
-  }
-
-  if (latestBlockCluster2.number) {
-    console.log(
-      `📍Trying to fetch block ${Number(
-        latestBlockCluster2.number
-      )} from cluster 1`
-    )
-    latestBlockCluster1 = await runCommand(
-      getBlock,
-      borHosts[0],
-      latestBlockCluster2.number,
-      maxRetries
-    )
-    if (latestBlockCluster1 === undefined) {
-      console.log(
-        `📍Unable to fetch block ${Number(
-          latestBlockCluster2.number
-        )} in cluster 1, exiting`
-      )
-      return
-    }
-
-    if (latestBlockCluster1.number) {
-      if (latestBlockCluster1.number !== latestBlockCluster2.number) {
-        console.log(
-          `📍Block number mismatch from clusters. Cluster 1: ${Number(
-            latestBlockCluster1.number
-          )}, Cluster 2: ${Number(latestBlockCluster2.number)}, exiting`
-        )
-        return
-      }
-
-      // Check if same block numbers have different hash or not
-      if (latestBlockCluster1.hash === latestBlockCluster2.hash) {
-        console.log(
-          `📍Block hash matched. Clusters are not created properly. Cluster 1: ${latestBlockCluster1.hash}, Cluster 2: ${latestBlockCluster2.hash}, exiting`
-        )
-        return
-      }
-
-      console.log(
-        `📍Same block found with different hash. Block number: ${Number(
-          latestBlockCluster1.number
-        )}, Cluster 1 hash: ${latestBlockCluster1.hash}, Cluster 2 hash: ${
-          latestBlockCluster2.hash
-        }`
-      )
-    } else {
-      console.log('📍Unable to fetch latest block from 1st cluster, exiting')
-      return
-    }
-  } else {
-    console.log('📍Unable to fetch latest block from 2nd cluster, exiting')
-    return
-  }
-
+  // Fetch same height blocks from different clusters and validate partition
+  let majorityForkBlock = await fetchSameHeightBlocks(borHosts[0], borHosts[2])
+ 
   // Expect no milestone to be proposed
   let latestMilestone = await fetchLatestMilestone(
     milestoneLength,
@@ -284,114 +117,15 @@ export async function milestonePartition() {
   await timer(4000)
 
   // Fetch block from cluster 2 to see if it got reorged to cluster 1
+  // Validate reorg by checking if cluster 2 got reorged to majority 
+  // fork i.e. cluster 1
+  await validateReorg(borHosts[2], majorityForkBlock)
   console.log(
-    `📍Fetching block ${Number(latestBlockCluster1.number)} from cluster 2`
+    '📍Cluster 2 successfully reorged to cluster 1 (having high difficulty) as expected'
   )
-  latestBlockCluster2 = await runCommand(
-    getBlock,
-    borHosts[2],
-    latestBlockCluster1.number,
-    maxRetries
-  )
-  if (latestBlockCluster2 === undefined) {
-    console.log(
-      `📍Unable to fetch block ${Number(
-        latestBlockCluster1.number
-      )} in cluster 2, exiting`
-    )
-    return
-  }
-
-  if (latestBlockCluster2.number) {
-    if (latestBlockCluster1.hash === latestBlockCluster2.hash) {
-      console.log(
-        '📍Cluster 2 successfully reorged to cluster 1 (having high difficulty) as expected'
-      )
-    } else {
-      console.log(
-        `📍Hash mismatch among clusters. Cluster 1 hash: ${latestBlockCluster1.hash}, Cluster 2 hash: ${latestBlockCluster2.hash}, exiting`
-      )
-      return
-    }
-  } else {
-    console.log(
-      `📍Unable to fetch ${Number(
-        latestBlockCluster1.number
-      )} in cluster 2, exiting`
-    )
-    return
-  }
 
   // For sanity check, also validate the latest block from both clusters
-  console.log(
-    '📍Fetching latest block from both cluster nodes for sanity check'
-  )
-  latestBlockCluster1 = await runCommand(
-    getBlock,
-    borHosts[0],
-    'latest',
-    maxRetries
-  )
-  if (latestBlockCluster1 === undefined) {
-    console.log('📍Unable to fetch latest block in cluster 1, exiting')
-    return
-  }
-
-  if (latestBlockCluster1.number) {
-    console.log(
-      `📍Trying to fetch block ${Number(
-        latestBlockCluster1.number
-      )} from cluster 2`
-    )
-    latestBlockCluster2 = await runCommand(
-      getBlock,
-      borHosts[2],
-      latestBlockCluster1.number,
-      maxRetries
-    )
-    if (latestBlockCluster2 === undefined) {
-      console.log(
-        `📍Unable to fetch block ${Number(
-          latestBlockCluster1.number
-        )} in cluster 2, exiting`
-      )
-      return
-    }
-
-    if (latestBlockCluster2.number) {
-      // check for block number
-      if (latestBlockCluster1.number !== latestBlockCluster2.number) {
-        console.log(
-          `📍Block number mismatch from clusters. Cluster 1: ${Number(
-            latestBlockCluster1.number
-          )}, Cluster 2: ${Number(latestBlockCluster2.number)}, exiting`
-        )
-        return
-      }
-
-      // check for block hash
-      if (latestBlockCluster1.hash !== latestBlockCluster2.hash) {
-        console.log(
-          `📍Block hash mismatch, failed reorg. Cluster 1: ${latestBlockCluster1.hash}, Cluster 2: ${latestBlockCluster2.hash}, exiting`
-        )
-        return
-      }
-
-      console.log(
-        `📍Same block found on both clusters. Block number: ${Number(
-          latestBlockCluster1.number
-        )}, Cluster 1 hash: ${latestBlockCluster1.hash}, Cluster 2 hash: ${
-          latestBlockCluster2.hash
-        }`
-      )
-    } else {
-      console.log('📍Unable to fetch latest block from 2nd cluster, exiting')
-      return
-    }
-  } else {
-    console.log('📍Unable to fetch latest block from 1st cluster, exiting')
-    return
-  }
+  await fetchAndValidateSameBlocks(borHosts[0], borHosts[2])
 
   // Wait for the next milestone to get proposed for verification
   latestMilestone = await fetchLatestMilestone(
